more decoupling and bug fix
